Add files via upload
So this outputs a graph for your file to be analysed 'pm' that shows how much the atmosphere sublimes throuhgout the year. It caps at 30% which is a bit of an issue but I'm not sure what to do about that. If the seasons don't overlap already in the model that output graph is ugly